design: 테마 경로 바꿈
